Set wgJsonDataNamespace default to [] instead of null to unbreak CI
All globals in extension.json have default=null, but the code expects them
to be arrays. This is causing the build to fail, as can be seen in the
mw-error.log file.

Note that this extension doesn't have a page on mw.org (!!!!!), so I'm
unsure what the right defaults are.

Change-Id: Ief973eace65b0cfc8a5e24b4c709bc6356a2cd9f