Make use of the more recent ?? operator
The main motivation here is that Phan actually complains about some
of these spots and suggests to use the ?? operator. I want to add
Phan in one of the next patches.

The code does the exact same as before. One nice benefit of the
modern syntax is that it avoids any kind of repetition.

Change-Id: I9c90a793b534c013f907022fe5b9e0ae1bfbd332