Fix all PHPCS issues related to PHPDocs
I'm quite sure about most of these types by now. How? Because I
already added Phan support to this codebase (see the next patch in
the chain) and Phan was actually able to help a lot.

I'm intentionally splitting this off as a separate patch. This patch
doesn't do anything but adding PHPDoc comments. As such it should
be impossible for this patch to break anything, even if one of the
types turns out to be not correct. It should be safe to merge this
any time. I hope this makes reviewing the later Phan patch much
easier.

Change-Id: I756e2d91eaad3f7065fde720593204d388e85318